fix tests
the client lib has been updated by introducing go mod files. this
required to change the test expectations / timeouts. the client lib
itself is only used in tests according to:

https://www.pivotaltracker.com/n/projects/2197709/stories/162818815

so it shouldn't be an issue that we pull in a slightly different version
than for the "legacy" 1.3.0 fork release.